refactor: inline config-callback logic into request-config.ts
