+++ src/lib/i18n/request-config.ts
@@ -1,3 +1,19 @@
-const { getRequestConfig } = require("next-intl/server")
-const { getI18NConfigCallback } = require("./config-callback")
-module.exports = getRequestConfig(getI18NConfigCallback)
+const { getI18NRequestConfig } = require("next-intl/server");
+const { fallbackLng, languages } = require("./settings");
+
+const requestConfig = getI18NRequestConfig(async ({ requestLocale }: { requestLocale: Promise<string | undefined> }) => {
+  // resolve the locale asynchronously
+  const resolvedRequestLocale = (await requestLocale) || fallbackLng;
+
+  // validate the locale and fallback if necessary
+  const resolvedLocale = languages.includes(resolvedRequestLocale)
+    ? resolvedRequestLocale
+    : fallbackLng;
+
+  return {
+    locale: resolvedLocale,
+    messages: (await import(`./translations/locales/${resolvedLocale}.ts`)).default,
+  };
+});
+
+export default requestConfig;
